+++ AddOns/Details/boot.lua
@@ -8,7 +8,7 @@
 
 		_detalhes.build_counter = 8516
 		_detalhes.alpha_build_counter = 8516 --if this is higher than the regular counter, use it instead
-		_detalhes.bcc_counter = 21
+		_detalhes.bcc_counter = 22
 		_detalhes.dont_open_news = true
 		_detalhes.game_version = version
 		_detalhes.userversion = version .. _detalhes.build_counter