+++ AddOns/LittleWigs/TBC/TheBlackMorass/Locales/ptBR.lua
@@ -5,6 +5,7 @@ if L then
 	L.wave_desc = "Anuncia mensagens de aviso da aproximação das ondas."
 
 	L.medivh = "Medivh"
+	L.rift = "Fenda Temporal"
 end
 
 L = BigWigs:NewBossLocale("Aeonus", "ptBR")