+++ AddOns/LittleWigs/TBC/TheMechanar/Locales/deDE.lua
@@ -18,4 +18,6 @@ L = BigWigs:NewBossLocale("Nethermancer Sepethrea", "deDE")
 if L then
 	L.inferno = "Inferno"
 	L.inferno_desc = "Tobende Flammen umgeben sich 8 Sek. lang mit einem Inferno, das allen Spielercharakteren in einem Umkreis von 10 Metern pro Sekunde Feuerschaden zufügt. Die tobenden Flammen können sich nicht bewegen oder Fähigkeiten einsetzen, solange sie in Flammen stehen. Wenn das Inferno endet, verfolgen sie ein neues Ziel."
+
+	L.fixate_desc = "Lässt den Zaubernden sich auf ein zufälliges Ziel fixieren."
 end