+++ AddOns/LittleWigs/TBC/TheMechanar/Locales/zhCN.lua
@@ -18,4 +18,6 @@ L = BigWigs:NewBossLocale("Nethermancer Sepethrea", "zhCN")
 if L then
 	L.inferno = "地狱火"
 	L.inferno_desc = "狂怒烈焰用地狱火环绕自己，持续8秒，每秒对10码范围内的所有玩家造成火焰伤害。燃烧时，狂怒烈焰不能移动或使用任何技能。地狱火效果消失时，狂怒烈焰会选择一个新目标去追逐。"
+
+	L.fixate_desc = "使施法者锁定一个随机目标。"
 end