+++ AddOns/BigWigs/CHANGELOG.md
@@ -1,7 +1,8 @@
 # BigWigs
 
-## [v26.2-classic](https://github.com/BigWigsMods/BigWigs/tree/v26.2-classic) (2021-06-08)
-[Full Changelog](https://github.com/BigWigsMods/BigWigs/compare/v26.1-classic...v26.2-classic) [Previous Releases](https://github.com/BigWigsMods/BigWigs/releases)
+## [v27-classic](https://github.com/BigWigsMods/BigWigs/tree/v27-classic) (2021-06-17)
+[Full Changelog](https://github.com/BigWigsMods/BigWigs/compare/v26.2-classic...v27-classic) [Previous Releases](https://github.com/BigWigsMods/BigWigs/releases)
 
+- Bump version  
+- Loader: Add warnings for installing the wrong version  
 - Sync with retail  
-- Fix some sound files